backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg[binary]>=3.1.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
